Refactor: Move tool directives to individual files in tool-directives directory
