+++ src/core/message-parsing/directives/index.ts
@@ -2,3 +2,6 @@ export type { Directive } from "./Directive"
 export type { TextDirective } from "./TextDirective"
 export type { ToolDirective } from "./ToolDirective"
 export type { LogDirective } from "./LogDirective"
+export type { ToolParamName, ToolResponse } from "./tool-directives"
+
+export { toolParamNames } from "./tool-directives"
