+++ src/core/message-parsing/directives/tool-directives/AskFollowupQuestionToolDirective.ts
@@ -1,0 +1,10 @@
+import { ToolDirective } from "../ToolDirective"
+import { ToolParamName } from "./ToolParamName"
+
+/**
+ * Directive for asking a follow-up question to the user.
+ */
+export interface AskFollowupQuestionToolDirective extends ToolDirective {
+	name: "ask_followup_question"
+	params: Partial<Pick<Record<ToolParamName, string>, "question" | "follow_up">>
+}
