+++ src/core/message-parsing/directives/tool-directives/FetchInstructionsToolDirective.ts
@@ -1,0 +1,10 @@
+import { ToolDirective } from "../ToolDirective"
+import { ToolParamName } from "./ToolParamName"
+
+/**
+ * Directive for fetching instructions to perform a task.
+ */
+export interface FetchInstructionsToolDirective extends ToolDirective {
+	name: "fetch_instructions"
+	params: Partial<Pick<Record<ToolParamName, string>, "task">>
+}
