+++ src/core/message-parsing/directives/tool-directives/ListFilesToolDirective.ts
@@ -1,0 +1,10 @@
+import { ToolDirective } from "../ToolDirective"
+import { ToolParamName } from "./ToolParamName"
+
+/**
+ * Directive for listing files and directories.
+ */
+export interface ListFilesToolDirective extends ToolDirective {
+	name: "list_files"
+	params: Partial<Pick<Record<ToolParamName, string>, "path" | "recursive">>
+}
