+++ src/core/message-parsing/directives/tool-directives/SearchFilesToolDirective.ts
@@ -1,0 +1,10 @@
+import { ToolDirective } from "../ToolDirective"
+import { ToolParamName } from "./ToolParamName"
+
+/**
+ * Directive for performing a regex search across files.
+ */
+export interface SearchFilesToolDirective extends ToolDirective {
+	name: "search_files"
+	params: Partial<Pick<Record<ToolParamName, string>, "path" | "regex" | "file_pattern">>
+}
