+++ src/core/message-parsing/directives/tool-directives/SwitchModeToolDirective.ts
@@ -1,0 +1,10 @@
+import { ToolDirective } from "../ToolDirective"
+import { ToolParamName } from "./ToolParamName"
+
+/**
+ * Directive for switching to a different mode.
+ */
+export interface SwitchModeToolDirective extends ToolDirective {
+	name: "switch_mode"
+	params: Partial<Pick<Record<ToolParamName, string>, "mode_slug" | "reason">>
+}
