+++ src/core/message-parsing/directives/tool-directives/WriteToFileToolDirective.ts
@@ -1,0 +1,10 @@
+import { ToolDirective } from "../ToolDirective"
+import { ToolParamName } from "./ToolParamName"
+
+/**
+ * Directive for writing content to a file.
+ */
+export interface WriteToFileToolDirective extends ToolDirective {
+	name: "write_to_file"
+	params: Partial<Pick<Record<ToolParamName, string>, "path" | "content" | "line_count">>
+}
