+++ src/core/message-parsing/presentAssistantMessage.ts
@@ -4,10 +4,9 @@ import { serializeError } from "serialize-error"
 import type { ToolName, ClineAsk, ToolProgressStatus } from "@roo-code/types"
 import { TelemetryService } from "@roo-code/telemetry"
 
-import type { LogDirective } from "./directives/LogDirective"
+import type { LogDirective, ToolParamName, ToolResponse } from "./directives"
 
 import { defaultModeSlug, getModeBySlug } from "../../shared/modes"
-import type { ToolParamName, ToolResponse } from "../../shared/tools"
 
 import { fetchInstructionsTool } from "../tools/fetchInstructionsTool"
 import { listFilesTool } from "../tools/listFilesTool"
