+++ src/core/tools/attemptCompletionTool.ts
@@ -4,7 +4,6 @@ import { TelemetryService } from "@roo-code/telemetry"
 
 import { Task } from "../task/Task"
 import {
-	ToolResponse,
 	AskApproval,
 	HandleError,
 	PushToolResult,
@@ -15,7 +14,7 @@ import {
 import { formatResponse } from "../prompts/responses"
 import { type ExecuteCommandOptions, executeCommand } from "./executeCommandTool"
 import { EXPERIMENT_IDS, experiments, experimentDefault } from "../../shared/experiments"
-import { ToolDirective } from "../message-parsing/directives"
+import { ToolDirective, ToolResponse } from "../message-parsing/directives"
 
 export async function attemptCompletionTool(
 	cline: Task,
